Separate overlay and menubar logic
