+++ InteractiveClassroom/ViewModel/Server/MenuBarViewModel.swift
@@ -2,43 +2,9 @@
 import SwiftUI
 import AppKit
 
-/// Handles macOS menu bar window management.
+/// Handles macOS menu bar window interactions.
 @MainActor
 final class MenuBarViewModel: ObservableObject {
-    private var overlayWindow: NSWindow?
-    private var originalPresentationOptions: NSApplication.PresentationOptions = []
-
-    /// Presents the full-screen overlay window.
-    func openOverlay(
-        pairingService: PairingService,
-        courseSessionService: CourseSessionService,
-        interactionService: InteractionService
-    ) {
-        closeOverlay()
-        originalPresentationOptions = NSApp.presentationOptions
-        NSApp.presentationOptions = originalPresentationOptions.union([.autoHideDock, .autoHideMenuBar])
-        let controller = NSHostingController(
-            rootView: ScreenOverlayView()
-                .environmentObject(pairingService)
-                .environmentObject(courseSessionService)
-                .environmentObject(interactionService)
-        )
-        let window = NSWindow(contentViewController: controller)
-        configureOverlayWindow(window)
-        window.makeKeyAndOrderFront(nil)
-        overlayWindow = window
-    }
-
-    /// Closes any existing overlay windows and restores normal presentation.
-    func closeOverlay() {
-        overlayWindow?.close()
-        let overlayWindows = NSApp.windows.filter { $0.identifier?.rawValue == "overlay" }
-        guard !overlayWindows.isEmpty || overlayWindow != nil else { return }
-        overlayWindows.forEach { $0.close() }
-        overlayWindow = nil
-        NSApp.presentationOptions = originalPresentationOptions
-    }
-
     /// Opens a window identified by `id` if one isn't already visible.
     /// If the window exists, it is brought to the front instead of creating a duplicate.
     func openWindowIfNeeded(id: String, openWindow: OpenWindowAction) {
@@ -47,22 +13,6 @@ final class MenuBarViewModel: ObservableObject {
         } else {
             openWindow(id: id)
         }
-    }
-
-    /// Applies identifier and screen configuration to the overlay window.
-    private func configureOverlayWindow(_ window: NSWindow) {
-        window.identifier = NSUserInterfaceItemIdentifier("overlay")
-        window.level = NSWindow.Level(rawValue: NSWindow.Level.mainMenu.rawValue - 1)
-        window.collectionBehavior = [.canJoinAllSpaces, .fullScreenAuxiliary, .stationary]
-        if let screenFrame = NSScreen.main?.frame {
-            window.setFrame(screenFrame, display: true)
-            window.contentView?.frame = screenFrame
-        }
-        window.styleMask = [.borderless]
-        window.isOpaque = false
-        window.backgroundColor = .clear
-        window.isReleasedWhenClosed = false
-        window.orderFrontRegardless()
     }
 }
 #endif
